requirements.txt
langchain
langchain-community
langgraph
openai
pyodbc
sqlalchemy
flask
python-dotenv